requirements.txt
pilkit==1.1.13
Pillow==3.1.0
